+++ lib/src/event/impl/dom_impl_standard.dart
@@ -44,11 +44,11 @@ class DomImplStandard extends DomImpl {
       child = child.nextNode; //nextSibling;
     }
 
-//    if (before == null) {
-//      parent.append(toAdd);
-//    } else {
+    if (before == null) {
+      parent.append(toAdd);
+    } else {
       parent.insertBefore(toAdd, before);
-//    }
+    }
   }
 
   dart_html.Element getChild(dart_html.Element elem, int index) {