+++ lib/src/ui/complex_panel.dart
@@ -246,12 +246,13 @@ abstract class ComplexPanel extends Panel implements IndexedPanelForIsWidget {
     // Physical attach.
     if (domInsert) {
       //DOM.insertChild(container, child.getElement(), beforeIndex);
-      dart_html.Element refChild = container.nodes[beforeIndex];
-      if (refChild != null) {
-        container.insertBefore(child.getElement(), refChild);
-      } else {
-        container.append(child.getElement());
-      }
+      Dom.insertChild(container, child.getElement(), beforeIndex);
+//      dart_html.Element refChild = container.nodes[beforeIndex];
+//      if (refChild != null) {
+//        container.insertBefore(child.getElement(), refChild);
+//      } else {
+//        container.append(child.getElement());
+//      }
     } else {
       //DOM.appendChild(container, child.getElement());
       container.append(child.getElement());